[rename] Recommendation inner class들 package 위치 inner로 변경
